+++ engine/table/src/test/java/io/deephaven/engine/table/impl/util/TestRedirectedColumnSource.java
@@ -33,7 +33,6 @@ import gnu.trove.list.TByteList;
 import gnu.trove.list.array.TByteArrayList;
 import io.deephaven.engine.testutil.junit4.EngineCleanup;
 import org.apache.commons.lang3.mutable.MutableInt;
-import org.apache.commons.lang3.mutable.MutableLong;
 import org.junit.Rule;
 import org.junit.Test;
 
@@ -73,7 +72,7 @@ public class TestRedirectedColumnSource {
             final RowSequence rs,
             final WritableObjectChunk<String, Values> chunk,
             final long offset) {
-        final MutableLong pos = new MutableLong();
+        final MutableInt pos = new MutableInt();
         rs.forAllRowKeys(k -> {
             final String s = (String) cs.get(k);
             assertEquals("offset=" + (pos.intValue() + offset) + ", k=" + k, s, chunk.get(pos.intValue()));
